main: change how NodeType is implemented
now node_id has a default implementation where it uses the TypeId of the node to identify the node, which is much safer and guarantees unique ids

+++ src/ast/src/nodeid.rs
@@ -7,7 +7,7 @@
 
 use std::fmt;
 
-type IdInner = u32;
+type IdInner = u64;
 
 /// An ID for each node type.
 ///
